Improve uncollider performance + add general PhysicsHelper
Until now, testing if a point was in a volume was done using a 0-sized
sphere, a physics world collision test and a callback result holder.
Performance was horrendous.

Since so far we only use polyhedral convex shapes, and
btPolyhedralConvexShape happens to have an isInside method, point
testing is now done using this (transforming said point with the inverse
of the btCollisionObject's world transform beforehand).

This directly calls the collision shape's penetration testing algorithm,
which bypasses any world-wise checks and callbacks, and that's what we
want.